refactor: 패키지 구조를 정리한다
