+++ android/app/src/main/java/com/digginroom/digginroom/data/di/ViewModelFactory.kt
@@ -7,8 +7,8 @@ import androidx.lifecycle.viewmodel.viewModelFactory
 import com.digginroom.digginroom.feature.join.JoinViewModel
 import com.digginroom.digginroom.feature.login.LoginViewModel
 import com.digginroom.digginroom.feature.room.RoomViewModel
-import com.digginroom.digginroom.feature.scrap.ScrapRoomViewModel
-import com.digginroom.digginroom.feature.scrap.ScrapViewModel
+import com.digginroom.digginroom.feature.scrap.viewmodel.ScrapRoomViewModel
+import com.digginroom.digginroom.feature.scrap.viewmodel.ScrapViewModel
 import com.digginroom.digginroom.model.RoomsModel
 
 class ViewModelFactory(context: Context) {
